refactor: optimize redux store
