+++ src/pages/login/index.jsx
@@ -3,14 +3,14 @@ import { useDispatch, useSelector } from "react-redux";
 import { Link, useLocation, useNavigate } from 'react-router-dom';
 import styles from './login.module.css';
 import { EmailInput, PasswordInput, Button } from '@ya.praktikum/react-developer-burger-ui-components';
-import { loginUser } from "../../services/actions/loginUser";
+import { loginUser } from "../../services/actions/userInfo";
 
 function Login() {
 
   const location = useLocation();
   const { from } = location.state || { from: { pathname: "/" } };
 
-  const { email: emailAfterReset } = useSelector((store) => store.resetPassword);
+  const { email: emailAfterReset } = useSelector((store) => store.userInfo);
 
   const dispatch = useDispatch();
   const navigate = useNavigate();
