+++ src/pages/register/index.jsx
@@ -3,7 +3,7 @@ import { Link, useNavigate } from 'react-router-dom';
 import styles from './register.module.css';
 import { EmailInput, PasswordInput, Input, Button } from '@ya.praktikum/react-developer-burger-ui-components';
 import { useDispatch } from "react-redux";
-import { registerUser } from "../../services/actions/registerUser";
+import { registerUser } from "../../services/actions/userInfo";
 
 function Register() {
